Adding WIP for totally unified handler.
