+++ configs/kinobi.cjs
@@ -463,6 +463,60 @@ kinobi.update(
         tokenStandard: { type: k.linkTypeNode("tokenStandard") },
       },
     },
+    print: {
+      accounts: {
+        editionMint: { isSigner: "either" },
+        masterTokenAccountOwner: { defaultsTo: k.identityDefault() },
+        editionTokenAccountOwner: { defaultsTo: k.identityDefault() },
+        editionMetadata: {
+          defaultsTo: k.pdaDefault("metadata", {
+            seeds: { mint: k.accountDefault("editionMint") },
+          }),
+        },
+        edition: {
+          defaultsTo: k.pdaDefault("masterEdition", {
+            seeds: { mint: k.accountDefault("editionMint") },
+          }),
+        },
+        editionTokenAccount: {
+          defaultsTo: ataPdaDefault("editionMint", "editionTokenAccountOwner"),
+        },
+        masterTokenAccount: {
+          defaultsTo: k.pdaDefault("associatedToken", {
+            importFrom: "mplToolbox",
+            seeds: {
+              mint: k.argDefault("masterEditionMint"),
+              owner: k.accountDefault("masterTokenAccountOwner"),
+            },
+          }),
+        },
+        masterMetadata: {
+          defaultsTo: k.pdaDefault("metadata", {
+            seeds: { mint: k.argDefault("masterEditionMint") },
+          }),
+        },
+        masterEdition: {
+          defaultsTo: k.pdaDefault("masterEdition", {
+            seeds: { mint: k.argDefault("masterEditionMint") },
+          }),
+        },
+        editionTokenRecord: {
+          defaultsTo: k.conditionalDefault("arg", "tokenStandard", {
+            value: k.vEnum("TokenStandard", "ProgrammableNonFungible"),
+            ifTrue: k.pdaDefault("tokenRecord", {
+              seeds: {
+                mint: k.accountDefault("editionMint"),
+                token: k.accountDefault("editionTokenAccount"),
+              },
+            }),
+          }),
+        },
+      },
+      args: {
+        masterEditionMint: { type: k.publicKeyTypeNode() },
+        tokenStandard: { type: k.linkTypeNode("tokenStandard") },
+      },
+    },
     updateMetadataAccountV2: {
       args: { updateAuthority: { name: "newUpdateAuthority" } },
     },
@@ -621,6 +675,27 @@ kinobi.update(
     verify: "verificationArgs",
     unverify: "verificationArgs",
   })
+);
+
+kinobi.update(
+  new k.TransformNodesVisitor([
+    {
+      selector: { kind: "instructionNode", name: "printV2" },
+      transformer: (node) => {
+        k.assertInstructionNode(node);
+        return k.instructionNode({
+          ...node,
+          accounts: [
+            ...node.accounts,
+            k.instructionAccountNode({
+              name: "holderDelegateRecord",
+              isOptional: true,
+              docs: ["The Delegate Record authorizing escrowless edition printing."],
+            })],
+        });
+      },
+    },
+  ])
 );
 
 // Update versioned instructions.
@@ -799,22 +874,6 @@ kinobi.update(
     },
     printV1: {
       accounts: {
-        editionMint: { isSigner: "either" },
-        editionMintAuthority: {
-          defaultsTo: k.accountDefault("masterTokenAccountOwner"),
-        },
-        masterTokenAccountOwner: { defaultsTo: k.identityDefault() },
-        editionTokenAccountOwner: { defaultsTo: k.identityDefault() },
-        editionMetadata: {
-          defaultsTo: k.pdaDefault("metadata", {
-            seeds: { mint: k.accountDefault("editionMint") },
-          }),
-        },
-        edition: {
-          defaultsTo: k.pdaDefault("masterEdition", {
-            seeds: { mint: k.accountDefault("editionMint") },
-          }),
-        },
         editionMarkerPda: {
           defaultsTo: k.conditionalDefault("arg", "tokenStandard", {
             value: k.vEnum("TokenStandard", "ProgrammableNonFungible"),
@@ -830,45 +889,37 @@ kinobi.update(
             }),
           }),
         },
-        editionTokenAccount: {
-          defaultsTo: ataPdaDefault("editionMint", "editionTokenAccountOwner"),
+        editionMintAuthority: {
+          defaultsTo: k.accountDefault("masterTokenAccountOwner"),
         },
-        masterTokenAccount: {
-          defaultsTo: k.pdaDefault("associatedToken", {
-            importFrom: "mplToolbox",
-            seeds: {
-              mint: k.argDefault("masterEditionMint"),
-              owner: k.accountDefault("masterTokenAccountOwner"),
-            },
-          }),
-        },
-        masterMetadata: {
-          defaultsTo: k.pdaDefault("metadata", {
-            seeds: { mint: k.argDefault("masterEditionMint") },
-          }),
-        },
-        masterEdition: {
-          defaultsTo: k.pdaDefault("masterEdition", {
-            seeds: { mint: k.argDefault("masterEditionMint") },
-          }),
-        },
-        editionTokenRecord: {
+      },
+      args: { edition: { name: "editionNumber" }, },
+    },
+    printV2: {
+      accounts: {
+        editionMarkerPda: {
           defaultsTo: k.conditionalDefault("arg", "tokenStandard", {
             value: k.vEnum("TokenStandard", "ProgrammableNonFungible"),
-            ifTrue: k.pdaDefault("tokenRecord", {
+            ifTrue: k.pdaDefault("editionMarkerV2", {
+              seeds: { mint: k.argDefault("masterEditionMint") },
+            }),
+            ifFalse: k.pdaDefault("editionMarkerFromEditionNumber", {
+              importFrom: "hooked",
               seeds: {
-                mint: k.accountDefault("editionMint"),
-                token: k.accountDefault("editionTokenAccount"),
+                mint: k.argDefault("masterEditionMint"),
+                editionNumber: k.argDefault("editionNumber"),
               },
             }),
           }),
         },
+        editionMintAuthority: {
+          defaultsTo: k.conditionalDefault("account", "holderDelegateRecord", {
+            ifTrue: k.accountDefault("payer"),
+            ifFalse: k.accountDefault("masterTokenAccountOwner"),
+          }),
+        },
       },
-      args: {
-        edition: { name: "editionNumber" },
-        masterEditionMint: { type: k.publicKeyTypeNode() },
-        tokenStandard: { type: k.linkTypeNode("tokenStandard") },
-      },
+      args: { edition: { name: "editionNumber" }, },
     },
     // Update.
     updateAsAuthorityItemDelegateV2:
